* ReSharper TidyUp

+++ Source/Krypton Components/Krypton.Toolkit/General/OutlookGrid/CustomColumns/DataGridViewPercentageCell.cs
@@ -72,7 +72,7 @@ namespace Krypton.Toolkit
             int barWidth;
             if ((double)value >= 1.0)
             {
-                barWidth = (int)(cellBounds.Width - 10);
+                barWidth = cellBounds.Width - 10;
             }
             else
             {
